feat: add TypeScript types and update posts page
