+++ client/src/app/types.ts
@@ -1,0 +1,12 @@
+interface User {
+    name: string;
+}
+
+interface Post {
+    id: number;
+    title: string;
+    content: string;
+    user: User;
+}
+
+export { User, Post };
